Login page

+++ src/main/java/com/example/bank/bank/model/Deposit.java
@@ -2,13 +2,13 @@ package com.example.bank.bank.model;
 
 public class Deposit {
 
-    int accountNumber;
+    long accountNumber;
     int amount;
     
-    public int getAccountNumber() {
+    public long getAccountNumber() {
         return accountNumber;
     }
-    public void setAccountNumber(int accountNumber) {
+    public void setAccountNumber(long accountNumber) {
         this.accountNumber = accountNumber;
     }
     public int getAmount() {
